+++ scripts/smoke-tests.sh
@@ -80,3 +80,4 @@ griffon service component-flaws npm
 griffon service component-flaws python-marshmallow --affectedness AFFECTED
 griffon service products-contain-component --search-all --search-upstreams -s libxml2 -a
 griffon service products-contain-component -s grep -v --search-all
+griffon service products-contain-component -r 'webkit.tk' -vv